==> 0. Merge_comments&metainfo.ipynb <==
# Proprocessing
This script is used for (1)merging raw datasets, (2) recode, and (3) split train and test sets:
- manual coding: *Data_FixedCodingMistakes_noICR_ReadyForAnalysis.csv*
- comments scraped from social media:
    - Youtube: *RE__cleaned_and_prepared_data/topic_youtube_comments_and_posts_ALL_strat_sample_orderbyMB_ReadyForCoding.xlsx* 
    and *youtube_general_subsample2_n850.xlsx*
    - Twitter: *Twitter_full-text-merged-with-comments_n300pergenre.xlsx*

The additional variables aggregated from manual coding:
- Incivility (general)
- Hate speech

16/08/2023
@Arya Liao

import pandas as pd
import numpy as np

## Import data

comment_info = ['ID','commentText','Mark_ID','showName','genre','Time_comment','likeCount_comment']
twitter_infoname = ['entities','user_info_Twitter','user_id','place','retweet_count','platform',
                    'retweeted','language','source','in_reply_to_status_id_str',
                    'in_reply_to_user_id_str','in_reply_to_screen_name','is_quote_status']
topic_youtube_infoname = ['videoTitle','description','Time_video','channelTitle','channelId','viewCount',
                          'dislikeCount_video','likeCount_video','date_difference',
                          'commentCount_video','replyCount_comment','topic','subscribers','language']
youtube_infoname = ['videoTitle', 'channelId','replyCount_comment','description','Time_video'] 

rename the inconsistent names

twitter_rename = {'id_str_x':'ID','tekst (of comment)':'commentText','twitter_name':'showName','created_at':'Time_comment',
                 'favorite_count':'likeCount_comment','user (check: name and screen_name)':'user_info_Twitter','lang':'language'}
topic_youtube_rename = {'commentId':'ID','publishDate':'Time_video','publishTime':'Time_comment','likeCount_x':'likeCount_video',
                       'likeCount_y':'likeCount_comment','dislikeCount':'dislikeCount_video','commentCount':'commentCount_video',
                       'replyCount':'replyCount_comment','topic_x':'topic'}
youtube_rename = {'commentId':'ID','publishDate':'Time_video','publishTime':'Time_comment','likeCount_y':'likeCount_comment',
                 'replyCount':'replyCount_comment'}

twitter = pd.read_excel('data/Raw_data/Twitter_full-text-merged-with-comments_n300pergenre.xlsx',dtype=str).rename(
    columns = twitter_rename).reset_index(drop=True)[comment_info+twitter_infoname]
topic_youtube = pd.read_excel('data/Raw_data/topic_youtube_comments_and_posts_ALL_strat_sample_orderbyMB_ReadyForCoding.xlsx',dtype=str).rename(
    columns = topic_youtube_rename).reset_index(drop=True)[comment_info+topic_youtube_infoname]
youtube = pd.read_excel('data/Raw_data/youtube_general_subsample2_n850.xlsx',dtype=str).rename(
    columns = youtube_rename).reset_index(drop=True)[comment_info+youtube_infoname]

combined = pd.concat([twitter,topic_youtube,youtube], ignore_index=True, sort=False)
combined = combined.drop_duplicates('ID')
combined

comments = pd.read_csv('data/Raw_data/Data_FixedCodingMistakes_noICR_ReadyForAnalysis.csv')

## Merge

whole = comments.merge(combined.drop(columns='Mark_ID'), how = 'left', on = 'ID')
whole.head()

### Fix missing comment text

missing_data = whole[whole['commentText'].isnull() == True]
missing_data[['ID','Mark_ID','Genre','topiccode','Platform','showName','genre']]

# comments = pd.read_csv('data/Raw_data/Data_FixedCodingMistakes_noICR_ReadyForAnalysis.csv')
for index,miss in missing_data.iterrows():
    ID = combined[combined['Mark_ID'] == str(miss['Mark_ID'])]
    comments.loc[index, 'ID'] = ID.loc[ID.index.tolist()[0],'ID']
    
whole = comments.merge(combined.drop(columns='Mark_ID'), how = 'left', on = 'ID')
whole.head()

missing_data = whole[whole['commentText'].isnull() == True]
missing_data[['ID','Mark_ID','Genre','topiccode','Platform','showName','genre']]

## Recoding

### aggregated variables

incivility = ['Name_calling', 'Vulgarity', 'Attack_reputation', 
                'Question_Intelligenc', 'All_caps_function', 'Sarcasm_to_criticize',
              'Individual_right', 'discrimination', 'Invoke_violence']
hate_list_focus = ['Name_calling', 'Vulgarity','Individual_right', 'discrimination', 'Invoke_violence']

for i, row in whole.iterrows():
    whole.loc[i, 'INCIVILITY_DUMMY'] = int(row[incivility].max()>0)
    whole.loc[i, 'HATELIST_FOCUSED_DUMMY'] = int(row[hate_list_focus].max()>0)
whole['HATELIST_FOCUSED_DUMMY'] = whole['HATELIST_FOCUSED_DUMMY'].astype(int)
whole['INCIVILITY_DUMMY'] = whole['INCIVILITY_DUMMY'].astype(int)


### recode date and hashtags

for index,row in whole.iterrows():
    whole.loc[index,'Time_comment_year'] = row['Time_comment'][0:4]
    try:
        whole.loc[index,'Time_video_year'] = row['Time_video'][0:4]
    except:
        pass
    try:
        hashtags = ast.literal_eval(row['entities'])['hashtags']
        if len(hashtags) > 0:
            whole.loc[index,'Hashtags'] = ''
            for hashtag in hashtags:
                whole.loc[index,'Hashtags'] = whole.loc[index,'Hashtags']+'#'+hashtag['text']
        whole.loc[index,'Source_clean'] = re.findall(r'Twitter .+</a>', row['source'])[0][8:-4]
    except:
        pass

whole

### Save to .csv and .dta

whole.to_csv('data/full_data.csv',index=False)

list_remove = []
for v in whole.columns:
    trydf = whole[[v,'ID']]
    try:
        trydf.to_stata('try.dta')
    except:
        list_remove.append(v)
list_remove.remove('ID')
list_remove

df4stata = whole.drop(columns = list_remove)
df4stata.to_stata('data/full_data.dta')

## Split

def get_split_index(number, ratio = 0.8):
    np.random.seed(123)
    train_index = np.random.choice(number, size = int(number*0.8), replace = False)
    train_index.sort()
    test_index = [i for i in range(number) if i not in train_index]
    return train_index, test_index

train_index, test_index = get_split_index(len(whole))
train_set = whole.loc[train_index,:]
test_set = whole.loc[test_index,:]

train_set.to_csv(f'data/train.csv',index=False)
test_set.to_csv(f'data/test.csv',index=False)

print(len(train_index), len(test_index))
print(test_index)
