==> 0.1 Merge_isreply.ipynb <==
import pandas as pd
import numpy as np
from sklearn.metrics import cohen_kappa_score, classification_report
import krippendorff

#open platform datasets:
twitter = pd.read_excel('data/Raw_data/Twitter_full-text-merged-with-comments_n300pergenre.xlsx',dtype=str)
topic_youtube = pd.read_excel('data/Raw_data/topic_youtube_comments_and_posts_ALL_strat_sample_orderbyMB_ReadyForCoding.xlsx',dtype=str)
youtube = pd.read_excel('data/Raw_data/youtube_general_subsample2_n850.xlsx',dtype=str)
comments = pd.read_csv('data/Raw_data/Data_FixedCodingMistakes_noICR_ReadyForAnalysis.csv')

comments.head()

comments.columns

pd.crosstab(comments['INTERACTIVITY_DUMMY'],comments['Interaction'])

pd.crosstab(comments['INTERACTIVITY_DUMMY'],comments['Acknowledgement'])

twitter.columns

#get full data:
df = pd.read_csv('data/Data_combination_test.csv')

#load LIWC 22 results:
liwc = pd.read_csv('outputs/automated_results/LIWC22_swear_result.csv')

liwc

#load train/test data
trainset: pd.DataFrame = pd.read_csv('data/train.csv')
testset: pd.DataFrame = pd.read_csv('data/test.csv')

#add train/test split variable to the dataset:
liwc.loc[liwc['Mark_ID'].isin(trainset['Mark_ID']), 'split'] = 'train'
liwc.loc[liwc['Mark_ID'].isin(testset['Mark_ID']), 'split'] = 'test'

#merge data:
df = df.merge(liwc, on='Mark_ID', how='left')

df.columns

pd.crosstab(df.mean_split_swear, df.split)

#evaluete respective performance on the TRAINING set (NOTE the wording of the few shot examples in the prompt is polished by GPT4o, not exactly like codebook):
llm_human_column_pairs = [
    ("INCIVILITY_DUMMY", "simple_swear"),
    ("INCIVILITY_DUMMY", "mean_split_swear"),
    ("INCIVILITY_DUMMY", "median_split_swear"),
    ("INCIVILITY_DUMMY", "q25_split_swear"),
    ("INCIVILITY_DUMMY", "q75_split_swear"),
    ("INCIVILITY_DUMMY", "q99_split_swear"),
   
]
print("Inter-coder reliablity: classification_report(human as gold, llm as predicted):")

for human_col, llm_col in llm_human_column_pairs:
    subset = df.loc[df.split == 'test', [human_col, llm_col]].dropna()
    human = subset[human_col].tolist()
    llm = subset[llm_col].tolist()

 

    print("---")
    print(f"{classification_report(human, llm)}")
