==> 1.2 Rationality_FK&fake_resource.ipynb <==
# Comparing dictionary and manual coding: Rationality
Data: *Data_FixedCodingMistakes_noICR_ReadyForAnalysis_with_comments*. 

This script will get rationality score based on the following methods:
1. Manually coding result: RATIONALITY_DUMMY
2. Flesch-Kincaid index via function `textstat.flesch_reading_ease`
3. Nithyanand et al. (2017b): a list of websites that is related to fake news
4. Munger and Phillips (2019): “alternative information networks”

import pandas as pd
import os
import textstat
import numpy as np
import re

### Read manual code and get the socre
only use colums about meta-information and about rationality

compare_rationality = pd.read_csv('data/full_data.csv')
compare_rationality.head()

len(compare_rationality)

Double check dummy variable calculation

def mannual_concrete(compare_rationality):
    for index, d in compare_rationality.iterrows():
        allsum = sum(d[['Reasoning','BackgroundInfo','ExternalEvidence']])
        compare_rationality.loc[index,'mannual_sum'] = allsum
        if allsum >= 1:
            compare_rationality.loc[index,'mannual_dummy'] = 1
        else:
            compare_rationality.loc[index,'mannual_dummy'] = 0
    return compare_rationality

compare_rationality = mannual_concrete(compare_rationality)

f = compare_rationality[['RATIONALITY_DUMMY','mannual_dummy']]
f.corr()

### Calculate FK score

def F_K_score(compare_rationality):
    for index, t in compare_rationality.iterrows():
        if type(t['commentText']) != float:
            f=textstat.flesch_reading_ease(str(t['commentText'])) 
            compare_rationality.loc[index,'FK_SCORE'] = f
        else:
            compare_rationality.loc[index,'FK_SCORE'] = np.nan
    return compare_rationality

compare_rationality = pd.DataFrame(F_K_score(compare_rationality))

### Find false sources

def detect_false_sources(text):
    urls = re.findall('[^ :/((www))]*?\.com',text)
    urls = urls+re.findall('[^ :/((www))]*?\.org',text)
    count = 0
    for url in urls:
        name = '.'.join(url.split('.')[-2:])
        if name in false_sources:
            print(repr(name))
            count +=1
    return count

def detect_ain(text):
    count = 0
    for a in ain:
        ainn = re.findall(a,text)
        if ainn:
            print(ainn)
            count += len(ain)
    return count

false_sources=pd.read_csv('https://raw.githubusercontent.com/BigMcLargeHuge/opensources/master/sources/sources.csv')["Unnamed: 0"].tolist()
ain = pd.read_csv('data/dictionaries/labeled_ain.csv')['title'].tolist()
print(len(false_sources),len(ain))

for index, text in compare_rationality.iterrows():
    compare_rationality.loc[index, 'FALSE_SCOURCES'] = 0
    if type(text['commentText']) == float:
        compare_rationality.loc[index, 'FALSE_SCOURCES'] = np.nan
        continue
    compare_rationality.loc[index, 'FALSE_SCOURCES'] = detect_false_sources(text['commentText'])
    compare_rationality.loc[index, 'FALSE_SCOURCES'] += detect_ain(text['commentText'])

As the result shows, only two comments contain false resources. So this methods may not sensitive in detecting rationality.

### Check descriptive result

compare_rationality.loc[compare_rationality['FK_SCORE']<10,['commentText','RATIONALITY_DUMMY','FK_SCORE']]

compare_rationality['FK_SCORE'].describe()

The reason FK score doesn't work may be that comments are short. The equation of [FK score](https://en.wikipedia.org/wiki/Flesch–Kincaid_readability_tests) is not suitable for short texts.

compare_rationality['RATIONALITY_DUMMY'].describe()

compare_rationality

compare_rationality.to_csv("outputs/automated_results/Rationality_FK&fake.csv",index=False)
