==> 1.3 Interactivity_dictionary.ipynb <==
# Measuring interactivity

import pandas as pd
import numpy as np
import re

### Read original CSV

#load data

data_path = 'data/full_data.csv'

interactivity_df = pd.read_csv(data_path) #import csv

#retain only important columns
interactivity_df_w_mention = interactivity_df[['ID','commentText']]


#create show list name
show_list = ["@TheDailyShow","@LastWeekTonight","@SouthPark","@nbcsnl","@colbertlateshow","@RealTimers","@TheOnion","@fullfrontalsamb",
"@JimmyKimmelLive","@LateNightSeth","@zondagmetlubach","@Lucky_TV","@klikbeet","@EvenTotHierBnnvara","@AC360","@TuckerCarlson","@hardball",
"@CBSEveningNews","@11thHour","@NewsHour","@ABCWorldNews","@Nightline","@FaceTheNation","60Minutes","@NBCNews","@MeetThePress","@NOS",
"@NOS","@nosop3","@RTLnieuws","@Nieuwsuur"] 

for show in show_list:
    interactivity_df_w_mention['commentText'] = interactivity_df_w_mention.apply(lambda x: x['commentText'].replace(show,""), axis = 1) #replace all appearing @ from the list

interactivity_df_w_mention['mentioned'] = interactivity_df_w_mention.apply(lambda x: 1 if "@" in x['commentText'] else 0, axis = 1) #return bool value whether contains @

interactivity_df_w_mention

### Checking presence of mention / @

#Subsetting Twitter 
interactivity_twi_w_mention = interactivity_df_w_mention[interactivity_df_w_mention['ID'].str.contains("^[0-9]",regex = True)]

#Subsetting YouTube
interactivity_yt_w_mention = interactivity_df_w_mention[interactivity_df_w_mention['ID'].str.contains("^Ug",regex = True)]

#drop commentText to avoid duplicated columns in the merge df
interactivity_twi_w_mention = interactivity_twi_w_mention[['ID','mentioned']] #twitter subset

interactivity_yt_w_mention = interactivity_yt_w_mention[['ID','mentioned']] #YT subset

interactivity_df_w_mention = interactivity_df_w_mention[['ID','mentioned']] #complete subset

interactivity_twi_w_mention 

### Merge and export

#merge Twitter subset
merge_with_interactivity_twi = interactivity_df.merge(interactivity_twi_w_mention, on = "ID")

#merge YouTube subset
merge_with_interactivity_yt = interactivity_df.merge(interactivity_yt_w_mention, on = "ID")

#merge complete
merge_with_interactivity = interactivity_df.merge(interactivity_df_w_mention, on = "ID")

merge_with_interactivity.to_csv("outputs/automated_results/interactivity.csv",index=False)

### Comparison with manually coded data

#Comparison Twitter
merge_with_interactivity_twi = merge_with_interactivity_twi[['ID','Interaction','Acknowledgement','mentioned']]
merge_with_interactivity_twi['manual_interactivity'] = merge_with_interactivity_twi.apply(lambda x: 1 if (x['Interaction']|x['Acknowledgement']) == 1 else 0, axis = 1)

#Comparison YouTube
merge_with_interactivity_yt = merge_with_interactivity_yt[['ID','Interaction','Acknowledgement','mentioned']]
merge_with_interactivity_yt['manual_interactivity'] = merge_with_interactivity_yt.apply(lambda x: 1 if (x['Interaction']|x['Acknowledgement']) == 1 else 0, axis = 1)

#Comparison complete
merge_with_interactivity = merge_with_interactivity[['ID','Interaction','Acknowledgement','mentioned']]
merge_with_interactivity['manual_interactivity'] = merge_with_interactivity.apply(lambda x: 1 if (x['Interaction']|x['Acknowledgement']) == 1 else 0, axis = 1)

#Correlation matrices for the subsets
import warnings

warnings.filterwarnings("ignore") #ignore deprecation wanrings 

print("Twitter:")
display(merge_with_interactivity_twi.corr()) # Twitter subset 
print("YouTube:")
display(merge_with_interactivity_yt.corr()) # YouTube subset
print("Complete:")
display(merge_with_interactivity.corr()) #complete set 

### Computing precision / recall

**Note.** The manual coded scores are now treated as actual score (aka golden standard)

def tp_fp_fn_tn(actual_score,predicted_score):
    #Let manual = actual score
    #TP: actual = 1, preducted = 1
    #FP: actual = 0, predicted = 1
    #FN: actual = 1, predicted = 0
    #TN: actual = 0, predicted = 0
    
    tp = sum((actual_score == 1) & (predicted_score ==1))
    fp = sum((actual_score == 0) & (predicted_score ==1))
    fn = sum((actual_score == 1) & (predicted_score ==0))
    tn = sum((actual_score == 0) & (predicted_score ==0))
    
    return tp, fp, fn, tn

#Precision and recall score 

def precision_score(tp, fp):
    precision = tp/(fp+tp)
    return precision

def recall_score(tp,fn):
    recall = tp/(fn+tp)
    
    return recall

#Calculate precision / recall for each dictionary
def precision_recall(dictionary):
    #calculate tp_fp_fn_tn for each dict

    tp, fp, fn, tn = tp_fp_fn_tn(merge_with_interactivity['manual_interactivity'], dictionary)
    
    precision = precision_score(tp,fp)
    recall = recall_score(tp,fn)
    
    return precision,recall

#Create PRF1 data

twitter_PR = precision_recall(merge_with_interactivity_twi['mentioned']) #twitter precision and recall
yt_PR = precision_recall(merge_with_interactivity_yt['mentioned']) #yt precision and recall

twitter_F1 = (twitter_PR[0] * twitter_PR[1]) / (twitter_PR[0] + twitter_PR[1]) #twitter F1 score
yt_F1 = (yt_PR[0] * yt_PR[1]) / (yt_PR[0] + yt_PR[1]) #yt F1 score

print(f"Precision, Recall & F1 of Twitter subset: \n {twitter_PR} {twitter_F1}")

print(f"Precision, Recall & F1 of YouTube subset: \n {yt_PR} {yt_F1}")
