==> 1.4 eMFD.ipynb <==
# Implementing Extended Moral Foundations Dictionary (eMFD) for rationality

Hopp, F. R., Fisher, J. T., Cornell, D., Huskey, R., & Weber, R. (2020). The extended Moral Foundations Dictionary (eMFD): Development and applications of a crowd-sourced approach to extracting moral intuitions from text. Behavior Research Methods, https://doi.org/10.3758/s13428-020-01433-0

## Please download Spacy, its English model and eMFDscore prior to the implementation
1. Install spacy (ver 3.4.0): *pip install Spacy* 
2. Install English model: *python -m spacy download en_core_web_sm* 
3. Install eMFDscore: *pip install https://github.com/medianeuroscience/emfdscore/archive/master.zip*

### Import dataset
We are going to use both full dataset and train set.

#Dependencies
import pandas as pd 
import numpy as np
from scipy import stats
import spacy

#import dataset
full_data = pd.read_csv(r"~\Desktop\Research\Social Media Analysis\publicsphere\data\sample\Data_ReadyForAnalysis_WithComments&MetaInfo.csv")
full_data = full_data[['commentText','ID']] #retain only comment text and comment ID
full_data = full_data.rename(columns={'commentText':0}) #need to be set to 0, otherwise will lead to keyerror 

### Implementation
As recommended by the authors, to discriminate which foundations are more or less represented in a text, we are going to use method 2:
1. Assign Single Probability per Word and Return Sentiment Scores

#Assign Single Probability per Word and Return Sentiment Scores
from emfdscore.scoring import score_docs 

num_docs = len(full_data) #select only comment text
 
DICT_TYPE = 'emfd' #use emfd as the dictionary
PROB_MAP = 'single' #assign a single probability to each word in the eMFD according to the foundation with the highest probability
SCORE_METHOD = 'bow' #bow approach (contextless)
OUT_METRICS = 'sentiment' #average sentiment for each foundation

df = score_docs(full_data,DICT_TYPE,PROB_MAP,SCORE_METHOD,OUT_METRICS,num_docs)

df

### Labelling conservative and liberals by likelihood

Conservatives and liberals are assumed to use distinct foundations while expressing opinions:
1. Liberals: Care and fairness
2. Conservatives: Loyalty, authority and sanctity

#Retain only useful columns
clean_df = df[['care_p','fairness_p','loyalty_p','authority_p','sanctity_p']]

#create max_likelihood_dict for storing likelihood and ideology
column_names = ['Liberal_foundation_likelihood', 'Conservative_foundation_likelihood', 'Likelihood_distance', 'Ideology']
max_likelihood_dict = pd.DataFrame(columns=column_names)


#Get max likelihood from both liberal and conservative foundations
max_likelihood_dict['Liberal_foundation_likelihood'] = clean_df[['care_p','fairness_p']].max(axis = 1) 
max_likelihood_dict['Conservative_foundation_likelihood'] = clean_df[['loyalty_p','authority_p','sanctity_p']].max(axis = 1)

#Get the likelihood distance between liberal and conservative
max_likelihood_dict['Likelihood_distance'] = max_likelihood_dict['Liberal_foundation_likelihood'] - max_likelihood_dict['Conservative_foundation_likelihood']

max_likelihood_dict['Ideology'] = max_likelihood_dict.apply(lambda x: 'Liberal' if x['Likelihood_distance'] >= 0 else 'Conservative', axis = 1)

max_likelihood_dict.head(5)

max_likelihood_dict

#Merge with original document
max_likelihood_dict = max_likelihood_dict.merge(full_data, left_index=True, right_index=True)

#rename columns
max_likelihood_dict = max_likelihood_dict.rename(columns = {0:'commentText'})

print("Full dataset:\n----------------------------------------------")
display(max_likelihood_dict)

#Create df with rows contained in train set
train_set = pd.read_csv(r"C:\Users\shrim\Desktop\Research\Social Media Analysis\publicsphere\data\sample\test_train_split\train.csv")
train_set = train_set[['ID']] #retain only comment ID

print("Train set subset:\n----------------------------------------------")
train_set_merge = train_set.merge(how = 'left',right = max_likelihood_dict, on='ID')
display(train_set_merge)

#export csv
max_likelihood_dict.to_csv('emfd_diversity.csv', index=False)
train_set_merge.to_csv('emfd_diversity_train_set.csv', index=False)
