==> 1.6 formality.ipynb <==
#Suppress all warnings
import warnings
warnings.filterwarnings("ignore")

import pandas as pd
import re
from nltk.stem import PorterStemmer
from nltk import pos_tag
from nltk import word_tokenize
import nltk

path = 'C:/Users/shrim/Desktop/Research/Social Media Analysis/publicsphere'

### Load comments content

df = pd.read_csv(f'{path}/data/sample/Data_ReadyForAnalysis_WithComments&MetaInfo.csv') # change file here
data = df[['ID','commentText']]

def processing_no_stemming(text):
    unstemmed = re.sub(' +',' ',' '.join([re.sub(r'http\S+|[^a-zA-Z0-9@]|\'', ' '
                                                 , w.lower()) for w in str(text).split()]).strip())
    return unstemmed

def processing_w_stemming(text):
    # define stemmer
    ps = PorterStemmer()
    stemmed = re.sub(' +',' ',' '.join([ps.stem(re.sub(r'http\S+|[^a-zA-Z0-9@]|\'',' '
                                               , w.lower())) for w in str(text).split()]).strip())
    return stemmed

#Stemming
data['CommentsStemmed'] = data['commentText'].apply(lambda row: processing_w_stemming(row))

# Produce another column with comments not stemmed, but processed:
data['CommentsNonStemmed'] = data['commentText'].apply(lambda row: processing_no_stemming(row))

### Tokenization and POS tagging

Please note that the POS are tagged by the universal tagset to align with the formality score equation.

#Tokenization
data['tokens'] = data['CommentsNonStemmed'].apply(lambda x: word_tokenize(x))

#POS tagging
data['POS'] = data['tokens'].apply(lambda x: pos_tag(x, tagset = "universal"))

### Calculating language formality


Calculating the formality scores using nouns, adjectives, articles and prepositions versus pronouns, adverbs, verbs and interjections.

The equation, using corresponding tags in the NLTK's universal POS Tagset, is as the following:

<font size = "3"><center> $\frac{{{NOUN}+{ADJ}+{DET}+{PRT}-{PRON}-{ADV}-{VERB}-{CONJ}}+100}{2}$ </center></font> 
(Heylighen & Dewaele, 2002)

#Count of POS per row
data['POS_count'] = data['POS'].apply(lambda x: nltk.FreqDist(tag for (word, tag) in x))

data['POS_count'].head(20)

#Calculating formality score
numerator = ['NOUN','ADJ','DET','PRT']
denominator = ['PRON','ADV','VERB','CONJ']

#Insert formality column with value 0 
data['formality'] = 0

for i in range(len(data)):
    num_temp = 0
    den_temp = 0
    
    for tag in range(3):
        num_temp += data['POS_count'][i][numerator[tag]]
        den_temp += data['POS_count'][i][denominator[tag]]
        
        formality_score = (num_temp - den_temp +100)/2
        
        data['formality'][i] = formality_score

### Comparing scores

Like other deliberative qualities, both dummy and sum rationality scores are used 

#Getting manually coded score
manual_rationality = df[['commentId','TopicRelevance','Reasoning','BackgroundInfo','ExternalEvidence','ExternalEvidence_1_TEXT']]

#create rationality score by summing the 4 indicators 
manual_rationality['rationality_score'] = manual_rationality.sum(axis = 1, numeric_only=True)

#create dummy variable
manual_rationality['dummy'] = manual_rationality['rationality_score'].apply(lambda x: 0 if x==0 else 1)

manual_rationality

#merge dataframes
data = data.merge(manual_rationality)
data = data.drop_duplicates(subset=['commentId'], ignore_index = True)

data

#Correlation matrix
data[['formality','rationality_score','dummy']].corr()
