==> 3. Combine_all_analysis.ipynb <==
## Combine all measures

This script combines all coding results from automated measures and generate a new dataset for evalution, including:
- Interactivity
    - @ mention
    - Machine learning
- Diversity
    - MFD 1.0, 2.0
    - eMFD
    - Machine learning
- Rationality
    - FK score
    - False sources
    - formality
    - integrative complexity
    - Machine learning
- Incivility
    - Six dictionaries
    - Machine learning


import pandas as pd

comments = pd.read_csv('data/test.csv')

## Read results of different measures

Define the filename and columns that will be included

#define variables
test_variables = ['INTERACTIVITY_DUMMY','INCIVILITY_DUMMY','HATELIST_FOCUSED_DUMMY',
                 'RATIONALITY_DUMMY','HAS_OPINION_DUMMY','LIBERAL_DUMMY','CONSERVATIVE_DUMMY']
variables_SML = [i+'_ML' for i in test_variables]
variables_FT = [i+'_FT' for i in test_variables]

files_path = 'outputs/automated_results/'
files_info = [['interactivity.csv',['mentioned']],
              ['prediction_ML.csv',variables_SML],
              ['prediction_FT.csv',variables_FT],
              ['new formality.csv',['formality']],
              ['formality_to_add.csv',['formality_dummy']],
              ['Rationality.csv',['FK_SCORE','FALSE_SCOURCES']],
              ['Integrative complexity.csv',['IC_score_per']],
              ['emfd_diversity.csv',['Liberal_foundation_likelihood','Conservative_foundation_likelihood','Likelihood_distance','Ideology']], #1=liberal, 0=conservative
              ['zero_shot_test_set.csv',['Civility_dummy','Rationality_dummy','Ideology_dummy','Interactivity_dummy']]
             ]

Merging all data

select = pd.read_csv('outputs/automated_results/Incivility&Diversity.csv')
df = comments.merge(select.drop(columns='commentText'), how = 'left', on = 'ID')

for filename, variables in files_info:
    new = pd.read_csv(files_path+filename)
    new = new.drop_duplicates('ID')
    df = df.merge(new[['ID']+variables], how = 'left', on = 'ID')
df

## Recode

### Recode FK score: 
- reverse the value: the higher the more complex
- make dummy variable
    - 0: lower than mean value = less complex
    - 1: higher than mean value = more complex

df['FK_SCORE'] = df['FK_SCORE'].max() - df['FK_SCORE']
df['FK_SCORE_DUMMY'] = 0
df.loc[df[df['FK_SCORE']>df['FK_SCORE'].mean()].index,'FK_SCORE_DUMMY'] = 1
df['FK_SCORE_DUMMY']

### Recode integrative complexity as dummy variable
- 0: higher than mean value = more complex
- 1: lower than mean value = less complex

df['Integrative_Complexity_dummy'] = 0
df.loc[df[df['IC_score_per']>df['IC_score_per'].mean()].index,'Integrative_Complexity_dummy'] = 1
df['Integrative_Complexity_dummy']

### Recode zeroshot results (removed from the paper)

df['Conservative_dummy_ZeroS'] = df['Ideology_dummy']
df['Liberal_dummy_ZeroS'] = 1 - df['Ideology_dummy']

### Recode eMFD/MFD based on likelihood
- Conservative
    - 1: liberal < conservative
    - 0: liberal >= conservative
- Liberal
    - 1: liberal > conservative
    - 0: liberal <= conservative

df['EMFD_Conservative_dummy'] = (df['Liberal_foundation_likelihood'] < df['Conservative_foundation_likelihood']).astype(int)
df['EMFD_Liberal_dummy'] = (df['Liberal_foundation_likelihood'] > df['Conservative_foundation_likelihood']).astype(int)

df['MFD1_liberal'] = (df['MFD1_liberal_ratio'] > df['MFD1_conservative_ratio']).astype(int)
df['MFD1_conservative'] = (df['MFD1_liberal_ratio'] < df['MFD1_conservative_ratio']).astype(int)

### Reverse Ksiazek's civility dictionary: the higher the more uncivil

df['Dict_Civility_Ksiazek_2015(reverse)'] = 1 - df['Dict_Civility_Ksiazek_2015']

## Save

df = df.rename(columns={'Ideology':'EMFD_ideology','Civility_dummy':'Civility_dummy_ZeroS','Rationality_dummy':'Rationality_dummy_ZeroS','Ideology_dummy':'Ideology_dummy_ZeroS','Interactivity_dummy':'Interactivity_dummy_ZeroS'})
df.to_csv('data/Data_combination_test.csv',index=False)
