==> 5.1 ML+Dic_Dictionaries_as_input_undersampling.ipynb <==
# Improvements: Dictionaries as input for machine learning

import sys,csv
from collections import Counter
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.naive_bayes import MultinomialNB
from sklearn.linear_model import LogisticRegression
from sklearn.tree import DecisionTreeClassifier
from sklearn.svm import SVC
from sklearn import metrics
from sklearn.pipeline import (make_pipeline, Pipeline)
from sklearn.metrics import make_scorer
from sklearn.model_selection import (train_test_split, cross_val_score, GridSearchCV)
from sklearn.utils import resample
import pandas as pd
import itertools

test_variables = ['INTERACTIVITY_DUMMY','INCIVILITY_DUMMY','HATELIST_FOCUSED_DUMMY',
                 'RATIONALITY_DUMMY','HAS_OPINION_DUMMY','LIBERAL_DUMMY','CONSERVATIVE_DUMMY']

dictionaries_variable = [[['INCIVILITY_DUMMY','HATELIST_FOCUSED_DUMMY'],
                          ['Dict_Hostility_Ksiazek_2015', 'Dict_Civility_Ksiazek_2015', 'Dict_GoogeProject_OffensiveWords', 
                         'Dict_Incivility_Muddiman', 'Dict_Swearwords_LIWC', 'Dict_HatebaseVocabEN']], 
                         [['HAS_OPINION_DUMMY','LIBERAL_DUMMY','CONSERVATIVE_DUMMY'],
                          ['MFD1_conservative_ratio', 'MFD1_liberal_ratio', 'MFD2_conservative_ratio', 'MFD2_liberal_ratio']]
                        ]

def down_sample_majority(df, majortopic):
        majority = int(len(df[df[majortopic]==0])/len(df)<0.5) # when the ratio of label=0 < .5, majority = 1, else majority = 0
        monority = 1 - majority # if majority = 1 then minority = 0, and vice versa
        df_majority = df[df[majortopic]==majority]
        df_minority = df[df[majortopic]==monority]
        df_majority_downsampled = resample(df_majority,
                                         replace=False,     #
                                         n_samples=len(df[df[majortopic]==1])) # set to N of minority topic
                                        # random_state=123) #

        df_downsampled = pd.concat([df_minority, df_majority_downsampled]).sample(frac=1)
        print(len(df_minority))
        return df_downsampled

def machine_learning(train, test, labels, dictionaries):
    acc = pd.DataFrame(columns = ['Classifier','Parameters', 'F1_score','Recall','Precision','Accuracy','Ratio_resampled'])

    df_downsampled = down_sample_majority(train, labels)
    train_labels = df_downsampled[labels]
    train_data = df_downsampled[dictionaries]
    test_labels = test[labels]
    test_data = test[dictionaries]
#    train_texts, test_texts, train_labels, test_labels = train_test_split(df_downsampled['commentText'].to_list(), df_downsampled[labels].to_list(), test_size=0.2, random_state=42)
#    print(f'after undersampling:\ntrain: {len(train_labels)}, test: {len(test_labels)}')
#    print(Counter(train_labels))
    grid = {"classifier__C": [0.01, 1, 100]}
    Classifiers = [MultinomialNB(), LogisticRegression(max_iter=1000),
                   SVC(kernel='rbf', class_weight="balanced"), SVC(kernel='linear', class_weight="balanced")
                  ]
    
    for classifier in Classifiers:
        pipeline = Pipeline(steps = [("classifier", classifier)])
        try:
            search=GridSearchCV(estimator=pipeline, n_jobs=-1, param_grid=grid, scoring='f1', cv=5)
            search.fit(train_data, train_labels)
        except:
            grid = {}
            search=GridSearchCV(estimator=pipeline, n_jobs=-1, param_grid=grid, scoring='f1', cv=5)
            search.fit(train_data, train_labels)
        #print(search.cv_results_['split1_test_score'])
        y_pred = search.predict(test_data)
        acc = acc.append({'Classifier':classifier,
                          'F1_score':metrics.f1_score(test_labels, y_pred),'Recall':metrics.recall_score(test_labels, y_pred),
                          'Precision':metrics.precision_score(test_labels, y_pred),'Accuracy':metrics.accuracy_score(test_labels, y_pred),
                          'Ratio_resampled':Counter(train_labels)[1]/(len(train_labels)),'Manual':test_labels,'Prediction':y_pred},ignore_index=True)
    best_classifier = acc[acc['F1_score'] == acc['F1_score'].max()].reset_index()
    #print('algorithm with maximum F1_score:', best_classifier)
    return acc, best_classifier['Prediction'][0]

dictionaries = pd.read_csv('outputs/automated_results/incivility&diversity.csv').drop(columns = 'commentText')

train_set = pd.read_csv('data/train.csv')[test_variables+['ID']+['commentText']]
train_set = train_set.merge(dictionaries, how = 'left', on = 'ID')

test_set = pd.read_csv('data/test.csv')[test_variables+['ID']+['commentText']]  
test_set = test_set.merge(dictionaries, how = 'left', on = 'ID')
train_set

accuracy = pd.DataFrame(columns = ['Variable', 'Classifier', 'F1_score','Recall','Precision','Accuracy','Ratio_test','Ratio_resampled','Manual','Prediction'])

for test_v, dictionaries in dictionaries_variable:
    for v in test_v:
        print(v)
        acc,prediction = machine_learning(train_set, test_set, v, dictionaries)
        test_set[v+'_ML'] = prediction

        acc['Variable'] = v
        acc['Ratio_test'] = test_set[v].mean()
        acc['Ratio_prediction'] = prediction.mean()
        accuracy = accuracy.append(acc,ignore_index=True)


accuracy

best_models = pd.DataFrame()
for v in test_variables:
    select_var = accuracy[accuracy['Variable'] == v]
    best_models = best_models.append(select_var[select_var['F1_score'] == select_var['F1_score'].max()],ignore_index=True)
best_models

## Update PRF1

from sklearn.metrics import classification_report

old = pd.read_csv('outputs/evaluation/PRF1_test.csv')

for index,data in best_models.iterrows():
    if data['Variable'] == 'HAS_OPINION_DUMMY':
        continue
    data['Manual'] = data['Manual'].astype(int)
    classification = classification_report(data['Manual'],data['Prediction'],output_dict=True)
    for c in ['macro avg','1','0']:
        old = old.append({'Dimension':'Incivility', 'Label':data['Variable'], 'Measures':'dict_as_input_for_ML','Class':c,'Precision':classification[c]['precision'],
                         'Recall':classification[c]['recall'],'F1-score':classification[c]['f1-score'],'support':classification[c]['support'],'Accuracy_overall':classification['accuracy']},ignore_index=True)

    print(data[['Variable','Classifier','Ratio_test']],'\n')


old.to_csv('outputs/evaluation/PRF1_test.csv',index=False)
