==> 5.3 DDR.ipynb <==
# Distributed Dictionary Representation (DDR)

Garten, J., Hoover, J., Johnson, K. M., Boghrati, R., Iskiwitch, C., & Dehghani, M. (2018). Dictionaries and distributions: Combining expert knowledge and large scale textual data content analysis: Distributed dictionary representation. Behavior research methods, 50, 344-361.

import os, sys
import pandas as pd

path = os.getcwd()
sys.path.insert(0, f"{path}/DDR/DDR-master")
import ddr

Download GoogleNews-vectors-negative300.bin <https://github.com/mmihaltz/word2vec-GoogleNews-vectors/tree/master>, and move it to the DDR folder before running the script

model, num_features, index2word_set = ddr.load_model(model_path = 'DDR/GoogleNews-vectors-negative300.bin')

## Run DDR model

dic_terms = ddr.load_terms.terms_from_txt(input_path = 'DDR/dictionaries')
agg_dic_vecs = ddr.get_vecs.dic_vecs(dic_terms = dic_terms,
                model = model,
                num_features = num_features,
                model_word_set = index2word_set)
ddr.load_terms.terms_to_csv(agg_dic_vecs, output_path = 'outputs/automated_results/Incivility&Diversity_ddr.csv', delimiter = ',')

agg_doc_vecs = ddr.get_vecs.doc_vecs_from_csv(input_path = 'data/test.csv',
            output_path = 'outputs/automated_results/ddr_result.csv',
            model = model,
            num_features = num_features,
            model_word_set = index2word_set,
            text_col = 'commentText',
            delimiter= ',',
            header = True,
            id_col = 'ID')


pd.read_csv('outputs/automated_results/Incivility&Diversity_ddr.csv').corr()

num_features = 300
ddr.get_loadings( agg_doc_vecs_path='outputs/automated_results/ddr_result.csv',
                 agg_dic_vecs_path='outputs/automated_results/Incivility&Diversity_ddr.csv',
                 out_path='outputs/automated_results/ddr_similarity.csv',
                 num_features=num_features,
                 delimiter = ',' )

Incivility = ['HATELIST_FOCUSED_DUMMY', 'INCIVILITY_DUMMY']
Diversity = ['HAS_OPINION_DUMMY','LIBERAL_DUMMY','CONSERVATIVE_DUMMY']

manual_coding = pd.read_csv('data/test.csv')
similarity = pd.read_csv('outputs/automated_results/ddr_similarity.csv')
similarity = similarity.merge(manual_coding[['ID','commentText']+Incivility+Diversity], on='ID', how='left')
similarity.to_csv('outputs/automated_results/ddr_merged_similarity.csv',index=False)
similarity

## Correlation

from scipy.stats import pearsonr
import numpy as np

rho = similarity.corr()
pval = similarity.corr(method=lambda x, y: pearsonr(x, y)[1]) - np.eye(*rho.shape)
p = pval.applymap(lambda x: ''.join(['*' for t in [0.001,0.01,0.05] if x<=t]))
matrix = rho.round(2).astype(str) + p
display(matrix)
